+++ src/connectors/zenodo/execution/zenodo_job_executor.py
@@ -45,7 +45,6 @@ from ...base.data.entities.connector import ConnectorJob
 from ...base.data.types import ProjectExternalStateCallbacks
 from ...base.execution import ConnectorJobExecutor
 from ...base.integration.execution.requests_executor import RequestsExecutorOptions
-from ...base.settings import TransmissionSettingIDs
 
 
 class ZenodoJobExecutor(ConnectorJobExecutor):